RavenDB-19953 Move logic to useIndexCleanup
