+++ src/Raven.Studio/typescript/commands/database/index/getIndexMergeSuggestionsCommand.ts
@@ -1,6 +1,7 @@
 ﻿import commandBase = require("commands/commandBase");
 import database = require("models/resources/database");
 import endpoints = require("endpoints");
+import IndexMergeResults = Raven.Server.Documents.Indexes.IndexMerging.IndexMergeResults;
 
 class getIndexMergeSuggestionsCommand extends commandBase {
 
@@ -11,9 +12,11 @@ class getIndexMergeSuggestionsCommand extends commandBase {
         this.db = db;
     }
 
-    execute(): JQueryPromise<Raven.Server.Documents.Indexes.IndexMerging.IndexMergeResults> {
+    execute(): JQueryPromise<IndexMergeResults> {
         const url = endpoints.databases.index.indexesSuggestIndexMerge;
-        return this.query(url, null, this.db);
+        
+        return this.query<IndexMergeResults>(url, null, this.db)
+            .fail((response: JQueryXHR) => this.reportError("Failed to get index merge suggestions", response.responseText, response.statusText));
     }
 }
